building and user public routes
